backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
Pillow==10.1.0
drf-yasg==1.21.7
PyJWT==2.8.0
cryptography==41.0.5
sqlparse==0.4.4
pytz==2023.3
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
djangorestframework-simplejwt[crypto]>=5.0.0,<6.0.0 
